project-backend/requirements.txt
Flask==2.1.0
Flask_Cors==3.0.10
Pillow==9.1.0
PyJWT==2.3.0
pytest==6.2.2
requests==2.25.1
Werkzeug==2.1.1